# Faster char validity checks