This PR speeds up `parse_literal` by over 90%; most of the speedup comes from reversing the order of operations in the loop + changing the checks from `&str`s to `char`s, and there's a little bit of extra boost from optimizing the number of comparison operations:
```
8 chars                 time:   [138.92 ns 138.98 ns 139.05 ns]
                        change: [-90.735% -90.716% -90.702%] (p = 0.00 < 0.05)

16 chars                time:   [216.19 ns 216.82 ns 217.52 ns]
                        change: [-91.568% -91.519% -91.479%] (p = 0.00 < 0.05)

32 chars                time:   [451.54 ns 452.01 ns 452.47 ns]
                        change: [-89.654% -89.579% -89.529%] (p = 0.00 < 0.05)

64 chars                time:   [722.19 ns 722.50 ns 722.83 ns]
                        change: [-90.617% -90.598% -90.579%] (p = 0.00 < 0.05)

128 chars               time:   [1.2490 µs 1.2528 µs 1.2575 µs]
                        change: [-91.389% -91.351% -91.314%] (p = 0.00 < 0.05)

256 chars               time:   [2.4217 µs 2.4246 µs 2.4278 µs]
                        change: [-91.546% -91.522% -91.501%] (p = 0.00 < 0.05)
```
Since `Sanitizer::parse_safe_char` was mentioned in the docs, I also compared this implementation to that one, and since the new one performed favorably:
```
8 chars                 time:   [145.44 ns 146.03 ns 146.69 ns]
                        change: [-2.7090% -2.4227% -2.1538%] (p = 0.00 < 0.05)

16 chars                time:   [215.61 ns 215.93 ns 216.25 ns]
                        change: [-9.8945% -9.4574% -9.0315%] (p = 0.00 < 0.05)

32 chars                time:   [439.28 ns 440.86 ns 443.31 ns]
                        change: [-3.5182% -2.8312% -1.9353%] (p = 0.00 < 0.05)

64 chars                time:   [713.74 ns 713.95 ns 714.18 ns]
                        change: [-6.0072% -5.8569% -5.7299%] (p = 0.00 < 0.05)

128 chars               time:   [1.2782 µs 1.2789 µs 1.2797 µs]
                        change: [-6.2688% -5.8473% -5.5859%] (p = 0.00 < 0.05)

256 chars               time:   [2.4301 µs 2.4314 µs 2.4330 µs]
                        change: [-5.6259% -5.2202% -4.7933%] (p = 0.00 < 0.05)
```
I applied it to `Sanitizer::parse_safe_char` as well.