@ljedrz 2022-10-07
Oh, and as for the relatively large bench times, those included some randomness; when it is removed (and a single 8192 random char string is used to ensure good distribution), the improvements are
- for the check in `parse_literal`: **-98.086%**
- for the check in `Parser::sanitize`: **-36.719%**