@ljedrz 2022-10-08
seems like a new nightly `clippy` just dropped :detective: